pyproject.toml
[project]
name = "Sharepoint"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "muhammadumairshaikh", email = "muhammadumair.shaikh@imperiumdynamics.com"},
]
dependencies = ["python-dotenv>=1.1.1", "msal>=1.34.0"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "python main.py"


[tool.pdm]
distribution = false
